One model per file. Remove deprecated key "active" in manifest file.
