+++ phonecall_plan/wizards/__init__.py
@@ -2,4 +2,4 @@
 # © 2016 6IT
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import phonecall_plan
+from . import add_call, add_partner_call
